remove authentication properties bean
